requirements.txt
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-readers-file
streamlit
watchfiles
